+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.3
+* Fixed contract verification in migration scripts
+* Updated batch length for EON loading phase
+
 ## 1.3.2
 * Added additional checks for signature format in VerificationLibrary
 * Added Testnet snapshot data and signatures